allocwatcher: Stat_t is unavailable on win
