+++ client/allocwatcher/alloc_watcher_test.go
@@ -5,19 +5,16 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strings"
-	"syscall"
 	"testing"
 	"time"
 
 	hclog "github.com/hashicorp/go-hclog"
 	"github.com/hashicorp/nomad/client/allocdir"
 	cstructs "github.com/hashicorp/nomad/client/structs"
-	ctestutil "github.com/hashicorp/nomad/client/testutil"
 	"github.com/hashicorp/nomad/helper/testlog"
 	"github.com/hashicorp/nomad/nomad/mock"
 	"github.com/hashicorp/nomad/nomad/structs"
@@ -199,148 +196,6 @@ func TestPrevAlloc_LocalPrevAlloc_Terminated(t *testing.T) {
 	// Since prev alloc is terminal, Wait should exit immediately with no
 	// context error
 	require.NoError(t, waiter.Wait(ctx))
-}
-
-// TestPrevAlloc_StreamAllocDir_Ok asserts that streaming a tar to an alloc dir
-// works.
-func TestPrevAlloc_StreamAllocDir_Ok(t *testing.T) {
-	ctestutil.RequireRoot(t)
-	t.Parallel()
-	dir, err := ioutil.TempDir("", "")
-	if err != nil {
-		t.Fatalf("err: %v", err)
-	}
-	defer os.RemoveAll(dir)
-
-	// Create foo/
-	fooDir := filepath.Join(dir, "foo")
-	if err := os.Mkdir(fooDir, 0777); err != nil {
-		t.Fatalf("err: %v", err)
-	}
-
-	// Change ownership of foo/ to test #3702 (any non-root user is fine)
-	const uid, gid = 1, 1
-	if err := os.Chown(fooDir, uid, gid); err != nil {
-		t.Fatalf("err : %v", err)
-	}
-
-	dirInfo, err := os.Stat(fooDir)
-	if err != nil {
-		t.Fatalf("err: %v", err)
-	}
-
-	// Create foo/bar
-	f, err := os.Create(filepath.Join(fooDir, "bar"))
-	if err != nil {
-		t.Fatalf("err: %v", err)
-	}
-	if _, err := f.WriteString("123"); err != nil {
-		t.Fatalf("err: %v", err)
-	}
-	if err := f.Chmod(0644); err != nil {
-		t.Fatalf("err: %v", err)
-	}
-	fInfo, err := f.Stat()
-	if err != nil {
-		t.Fatalf("err: %v", err)
-	}
-	f.Close()
-
-	// Create foo/baz -> bar symlink
-	if err := os.Symlink("bar", filepath.Join(dir, "foo", "baz")); err != nil {
-		t.Fatalf("err: %v", err)
-	}
-	linkInfo, err := os.Lstat(filepath.Join(dir, "foo", "baz"))
-	if err != nil {
-		t.Fatalf("err: %v", err)
-	}
-
-	buf := new(bytes.Buffer)
-	tw := tar.NewWriter(buf)
-
-	walkFn := func(path string, fileInfo os.FileInfo, err error) error {
-		// Include the path of the file name relative to the alloc dir
-		// so that we can put the files in the right directories
-		link := ""
-		if fileInfo.Mode()&os.ModeSymlink != 0 {
-			target, err := os.Readlink(path)
-			if err != nil {
-				return fmt.Errorf("error reading symlink: %v", err)
-			}
-			link = target
-		}
-		hdr, err := tar.FileInfoHeader(fileInfo, link)
-		if err != nil {
-			return fmt.Errorf("error creating file header: %v", err)
-		}
-		hdr.Name = fileInfo.Name()
-		tw.WriteHeader(hdr)
-
-		// If it's a directory or symlink we just write the header into the tar
-		if fileInfo.IsDir() || (fileInfo.Mode()&os.ModeSymlink != 0) {
-			return nil
-		}
-
-		// Write the file into the archive
-		file, err := os.Open(path)
-		if err != nil {
-			return err
-		}
-		defer file.Close()
-
-		if _, err := io.Copy(tw, file); err != nil {
-			return err
-		}
-
-		return nil
-	}
-
-	if err := filepath.Walk(dir, walkFn); err != nil {
-		t.Fatalf("err: %v", err)
-	}
-	tw.Close()
-
-	dir1, err := ioutil.TempDir("", "nomadtest-")
-	if err != nil {
-		t.Fatalf("err: %v", err)
-	}
-	defer os.RemoveAll(dir1)
-
-	rc := ioutil.NopCloser(buf)
-	prevAlloc := &remotePrevAlloc{logger: testlog.HCLogger(t)}
-	if err := prevAlloc.streamAllocDir(context.Background(), rc, dir1); err != nil {
-		t.Fatalf("err: %v", err)
-	}
-
-	// Ensure foo is present
-	fi, err := os.Stat(filepath.Join(dir1, "foo"))
-	if err != nil {
-		t.Fatalf("err: %v", err)
-	}
-	if fi.Mode() != dirInfo.Mode() {
-		t.Fatalf("mode: %v", fi.Mode())
-	}
-	stat := fi.Sys().(*syscall.Stat_t)
-	if stat.Uid != uid || stat.Gid != gid {
-		t.Fatalf("foo/ has incorrect ownership: expected %d:%d found %d:%d",
-			uid, gid, stat.Uid, stat.Gid)
-	}
-
-	fi1, err := os.Stat(filepath.Join(dir1, "bar"))
-	if err != nil {
-		t.Fatalf("err: %v", err)
-	}
-	if fi1.Mode() != fInfo.Mode() {
-		t.Fatalf("mode: %v", fi1.Mode())
-	}
-
-	fi2, err := os.Lstat(filepath.Join(dir1, "baz"))
-	if err != nil {
-		t.Fatalf("err: %v", err)
-	}
-	if fi2.Mode() != linkInfo.Mode() {
-		t.Fatalf("mode: %v", fi2.Mode())
-	}
 }
 
 // TestPrevAlloc_StreamAllocDir_Error asserts that errors encountered while
